Allow skipping the build of hfsutils
I've been working on putting together a Homebrew formula in my own tap to make installing Retro68 simple, and due to the nature of extracting classic disk images like MPW-GM on a modern Mac, I've ended up needing to depend on hfsutils to prep the libraries. This then means that the Homebrew formula of Retro68 won't successfully install because Homebrew already sees the hfsutils programs in the place it's expecting to put them.

This patch adds a `--skip-hfsutils` flag to build-toolchain.bash, which skips only the build of hfsutils. This way we can just use the Homebrew-provided version (which is identical anyway) and Retro68 no longer conflicts with it.

I've already been building Retro68 on my machine with this patch, so I know it builds at least once, but I'm not 100% sure I've added the right conditional to rebuild checks, so feedback is very welcome.